# Optimize iterators a bit