--- a/Jint/Native/Array/ArrayPrototype.cs
+++ b/Jint/Native/Array/ArrayPrototype.cs
@@ -1273,7 +1273,7 @@ namespace Jint.Native.Array
             else
             {
                 var elementObj = TypeConverter.ToObject(_realm, firstElement);
-                var func = elementObj.Get("toLocaleString", elementObj) as ICallable;
+                var func = elementObj.Get("toLocaleString") as ICallable;
                 if (func is null)
                 {
                     ExceptionHelper.ThrowTypeError(_realm);
@@ -1292,7 +1292,7 @@ namespace Jint.Native.Array
                 else
                 {
                     var elementObj = TypeConverter.ToObject(_realm, nextElement);
-                    var func = elementObj.Get("toLocaleString", elementObj) as ICallable;
+                    var func = elementObj.Get("toLocaleString") as ICallable;
                     if (func is null)
                     {
                         ExceptionHelper.ThrowTypeError(_realm);