* use same type for all iterators which has faster shortcuts
* add fast path for array get length
* small cleanup

## Jint.Benchmark.ForBencmark

| **Diff**|Method|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |ForOfVar|164.7 μs|0.35 μs|330.99 KB|
| **New** |	| **107.7 μs (-35%)** | **0.22 μs** | **119.95 KB (-64%)** |
| Old |ForOfLet|177.4 μs|0.64 μs|635.59 KB|
| **New** |	| **139.8 μs (-21%)** | **0.30 μs** | **424.55 KB (-33%)** |
| Old |ForOfConst|178.6 μs|0.45 μs|635.59 KB|
| **New** |	| **144.3 μs (-19%)** | **0.38 μs** | **424.55 KB (-33%)** |
| Old |ForVar|159.9 μs|0.67 μs|3.66 KB|
| **New** |	| **152.4 μs (-5%)** | **1.62 μs** | **3.66 KB (0%)** |
| Old |ForLet|210.1 μs|0.44 μs|191.42 KB|
| **New** |	| **211.8 μs (+1%)** | **3.48 μs** | **191.42 KB (0%)** |